Removed comments

+++ Source/Camera/Camera.cpp
@@ -68,8 +68,8 @@ void Camera::update()
 {
 
 	XMVECTOR move;
-	move = XMVectorSet(0.1, 0.1, 0.1, 0.0);
-	m_eye -= move * Timer::getInstance()->getDt();
+	move = XMVectorSet(0.1f, 0.1f, 0.1f, 0.0f);
+	m_eye -= move * (float)Timer::getInstance()->getDt();
 
 	m_matrices.m_view = XMMatrixLookAtRH(m_eye, m_target, m_up);
 	m_matrices.m_proj = XMMatrixPerspectiveFovRH(m_fov, m_aspectRatio, m_nearPlane, m_farPlane);
